requirements.txt
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.31.0
InquirerPy==0.3.4